// src/app/core/interfaces/interface.user.ts
export interface UserFull extends User {
  totalPurchase: number;
  userFullName: string;
  profileImage?: string;
}

export interface User {
  address: {
    geolocation: {
      lat: number;
      long: number;
    };
    city: string;
    street: string;
    number: number;
    zipcode: string;
  }
  id: number;
  email: string;
  username: string;
  password: string;
  name: {
    firstname: string;
    lastname: string;
  };
  phone: string;
  __v: number
}